pyproject.toml
[tool.poetry]
name = "flask-app-db"
version = "0.1.0"
description = ""
authors = ["AndreyBerezyanskiy <andrey.berezyanskiy@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_app_db"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-wtf = "^1.1.1"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
faker = "^18.10.0"
psycopg2-binary = "^2.9.6"
flask-login = "^0.6.2"
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
